requirements.txt
Django>=4.0,<5.0
firebase-admin>=6.0,<7.0
djangorestframework>=3.12,<4.0
drf-spectacular>=0.20,<1.0
